travis: Bump wolfSSL to 4.4.0
SHA-3 is only automatically enabled on x86/x64.  The tests are disabled
because we don't need them and they cause a compile warning/error when
built with clang on x64 (sizeof() on a pointer to an array).  If the
examples are enabled another test suite is built, which includes the
disabled crypto tests.